requirements.txt
ollama
streamlit
pypdf2
sentence-transformers